[#202] chore: fonts 디렉토리 이름 변경
